Nose (by default) does not allow files called "setup.py" to be imported
